20230325 airnode contract addressess (#352)
* Scripts for axios files
Rather run axios at runtime we now use a script and build local files. This enables the files to be render at build time into the html pages, thus they can now be indexed by flex.

* update flex indexes
